# docs/swagger.yaml
definitions:
...
  hotel.Hotel:
    properties:
      city:
        type: string
      id:
        type: string
      name:
        type: string
      stars:
        type: integer
    type: object
...
  httputils.ErrorResponse:
    properties:
      error:
        example: failed to parse query value 'stars-from'
        type: string
      message:
        example: failed to find hotels
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputils.SuccessResponse:
    properties:
      data: {}
      message:
        example: successfully found hotels
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/hotel/create:
    post:
      consumes:
      - application/json
      description: Создаёт отель с переданными параметрами
      parameters:
      - description: Город
        in: query
        name: city
        required: true
        type: string
      - description: Количество звёзд
        in: query
        name: stars
        required: true
        type: integer
      - description: Название отеля (необязатено)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/hotel.Hotel'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Создать отель (Admin only)
      tags:
      - Отель
  /api/admin/hotel/delete:
    delete:
      consumes:
      - application/json
      description: Удаляет отель по ID
      parameters:
      - description: ID отеля
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Удалить отель (Admin only)
      tags:
      - Отель
  /api/admin/hotel/update:
    put:
      consumes:
      - application/json
      description: Изменяет отель переданными параметрами
      parameters:
      - description: ID отеля
        in: query
        name: id
        required: true
        type: string
      - description: Новый город (необязатено)
        in: query
        name: city
        type: string
      - description: Новое количество звёзд (необязатено)
        in: query
        name: stars
        type: integer
      - description: Новое название отеля (необязатено)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/hotel.Hotel'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Изменить отель (Admin only)
      tags:
      - Отель
...
  /api/hotel/find:
    get:
      consumes:
      - application/json
      description: Находит отель по фильтрам. Можно искать без фильтров. Нельзя использовать
        одновременно 'stars' и 'stars-from' + 'stars-to'. Если указать 'arrange',
        то билеты будут отсортированы по цене в указанном порядке (asc - по возрастанию,
        desc - по убыванию)
      parameters:
      - description: ID отеля (необязатено)
        in: query
        name: id
        type: string
      - description: Город (необязатено)
        in: query
        name: city
        type: string
      - description: Количество звёзд (необязатено)
        in: query
        name: stars
        type: integer
      - description: Минимальное кол-во звезд (необязатено)
        in: query
        name: stars-from
        type: integer
      - description: Максимальное кол-во звезд (необязатено)
        in: query
        name: stars-to
        type: integer
      - description: Упорядочить по звездам ('asc' возрастание, 'desc' убывание) (необязатено)
        in: query
        name: arrange
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/hotel.Hotel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Найти отель по параметрам
      tags:
      - Отель
...
swagger: "2.0"